Added STATS, tested them.

+++ test_storer.sh
@@ -73,6 +73,35 @@ echo -e '\e[1;32mOK\e[0m'
 rm -rf $TMPDIR
 
 
+echo -n 'Writes an entry and keeps stats: '
+mkdir -p $TMPDIR
+cat >$TMPDIR/input.txt <<EOF
+{"data": 42, "ms":0}
+UNITTEST_STATS
+EOF
+cat >$TMPDIR/golden.txt <<EOF
+{"data":42,"ms":0}
+EOF
+cat >$TMPDIR/golden_output.txt <<EOF
+{"total_consumed":1,"total_replayed":0,"total_file_renames":0,"current_file":{"entries":1,"ms":{"earliest":0,"latest":0}}}
+EOF
+cat $TMPDIR/input.txt | $BINARY --storer_max_time_discrepancy_ms=1e15 >$TMPDIR/output.txt
+if [ $(ls $TMPDIR/destination/* | wc -l) != 1 ] ; then
+    echo -e '\e[1;31mFAIL\e[0m'
+    exit 1
+fi
+if ! cat $TMPDIR/destination/* | sort | $DIFF - $TMPDIR/golden.txt ; then
+    echo -e '\e[1;31mFAIL\e[0m'
+    exit 1
+fi
+if ! cat $TMPDIR/output.txt | grep "^UNITTEST" | cut -f 2 | $DIFF - $TMPDIR/golden_output.txt ; then
+    echo -e '\e[1;31mFAIL\e[0m'
+    exit 1
+fi
+echo -e '\e[1;32mOK\e[0m'
+rm -rf $TMPDIR
+
+
 echo -n 'Fails if intermediate directory can not be written into: '
 mkdir -p $TMPDIR
 mkdir -p $TMPDIR/intermediate
@@ -149,12 +178,20 @@ echo -n 'Explicitly flushes: '
 mkdir -p $TMPDIR
 cat >$TMPDIR/input.txt <<EOF
 {"ms":0,"data":"1foo"}
+UNITTEST_STATS
 FLUSH
-{"ms":0,"data":"2bar"}
+UNITTEST_STATS
+{"ms":1,"data":"2bar"}
+UNITTEST_STATS
 EOF
 cat >$TMPDIR/golden.txt <<EOF
 {"ms":0,"data":"1foo"}
-{"ms":0,"data":"2bar"}
+{"ms":1,"data":"2bar"}
+EOF
+cat >$TMPDIR/golden_output.txt <<EOF
+{"total_consumed":1,"total_replayed":0,"total_file_renames":0,"current_file":{"entries":1,"ms":{"earliest":0,"latest":0}}}
+{"total_consumed":1,"total_replayed":0,"total_file_renames":1,"current_file":{}}
+{"total_consumed":2,"total_replayed":0,"total_file_renames":1,"current_file":{"entries":1,"ms":{"earliest":1,"latest":1}}}
 EOF
 cat $TMPDIR/input.txt | $BINARY --storer_max_time_discrepancy_ms=1e15 >$TMPDIR/output.txt
 if [ $(ls $TMPDIR/destination/* | wc -l) != 2 ] ; then
@@ -165,16 +202,25 @@ if ! cat $TMPDIR/destination/* | sort | $DIFF - $TMPDIR/golden.txt ; then
     echo -e '\e[1;31mFAIL\e[0m'
     exit 1
 fi
+if ! cat $TMPDIR/output.txt | grep "^UNITTEST" | cut -f 2 | $DIFF - $TMPDIR/golden_output.txt ; then
+    echo -e '\e[1;31mFAIL\e[0m'
+    exit 1
+fi
 echo -e '\e[1;32mOK\e[0m'
 rm -rf $TMPDIR
 
 echo -n 'Implicitly flushes by maximum number of entires per file: '
 mkdir -p $TMPDIR
 cat >$TMPDIR/input.txt <<EOF
+UNITTEST_STATS
 {"ms":0,"data":"1foo1"}
+UNITTEST_STATS
 {"ms":0,"data":"2bar1"}
+UNITTEST_STATS
 {"ms":0,"data":"1foo2"}
+UNITTEST_STATS
 {"ms":0,"data":"2bar2"}
+UNITTEST_STATS
 EOF
 cat >$TMPDIR/golden.txt <<EOF
 {"ms":0,"data":"1foo1"}
@@ -182,12 +228,23 @@ cat >$TMPDIR/golden.txt <<EOF
 {"ms":0,"data":"2bar1"}
 {"ms":0,"data":"2bar2"}
 EOF
+cat >$TMPDIR/golden_output.txt <<EOF
+{"total_consumed":0,"total_replayed":0,"total_file_renames":0,"current_file":{"entries":0,"ms":{"earliest":null,"latest":null}}}
+{"total_consumed":1,"total_replayed":0,"total_file_renames":0,"current_file":{"entries":1,"ms":{"earliest":0,"latest":0}}}
+{"total_consumed":2,"total_replayed":0,"total_file_renames":0,"current_file":{"entries":2,"ms":{"earliest":0,"latest":0}}}
+{"total_consumed":3,"total_replayed":0,"total_file_renames":1,"current_file":{"entries":0,"ms":{"earliest":null,"latest":null}}}
+{"total_consumed":4,"total_replayed":0,"total_file_renames":1,"current_file":{"entries":1,"ms":{"earliest":0,"latest":0}}}
+EOF
 cat $TMPDIR/input.txt | $BINARY --storer_max_time_discrepancy_ms=1e15 --storer_max_entries_per_file=2 >$TMPDIR/output.txt
 if [ $(ls $TMPDIR/destination/* | wc -l) != 2 ] ; then
     echo -e '\e[1;31mFAIL\e[0m'
     exit 1
 fi
 if ! cat $TMPDIR/destination/* | sort | $DIFF - $TMPDIR/golden.txt ; then
+    echo -e '\e[1;31mFAIL\e[0m'
+    exit 1
+fi
+if ! cat $TMPDIR/output.txt | grep "^UNITTEST" | cut -f 2 | $DIFF - $TMPDIR/golden_output.txt ; then
     echo -e '\e[1;31mFAIL\e[0m'
     exit 1
 fi
@@ -233,13 +290,21 @@ echo -n 'Implicitly flushes because of the timeout (flaky, uses sleep): '
 mkdir -p $TMPDIR
 cat >$TMPDIR/i1.txt <<EOF
 {"ms":0,"data":"foo"}
+UNITTEST_STATS
 EOF
 cat >$TMPDIR/i2.txt <<EOF
-{"ms":0,"data":"bar"}
+UNITTEST_STATS
+{"ms":1,"data":"bar"}
+UNITTEST_STATS
 EOF
 cat >$TMPDIR/golden.txt <<EOF
-{"ms":0,"data":"bar"}
 {"ms":0,"data":"foo"}
+{"ms":1,"data":"bar"}
+EOF
+cat >$TMPDIR/golden_output.txt <<EOF
+{"total_consumed":1,"total_replayed":0,"total_file_renames":0,"current_file":{"entries":1,"ms":{"earliest":0,"latest":0}}}
+{"total_consumed":1,"total_replayed":0,"total_file_renames":1,"current_file":{"entries":0,"ms":{"earliest":null,"latest":null}}}
+{"total_consumed":2,"total_replayed":0,"total_file_renames":1,"current_file":{"entries":1,"ms":{"earliest":1,"latest":1}}}
 EOF
 (cat $TMPDIR/i1.txt ; sleep 1 ; cat $TMPDIR/i2.txt) | $BINARY --storer_max_time_discrepancy_ms=1e15 --storer_max_file_age_ms=200 >$TMPDIR/output.txt
 if [ $(ls $TMPDIR/destination/* | wc -l) != 2 ] ; then
@@ -247,6 +312,10 @@ if [ $(ls $TMPDIR/destination/* | wc -l) != 2 ] ; then
     exit 1
 fi
 if ! cat $TMPDIR/destination/* | sort | $DIFF - $TMPDIR/golden.txt ; then
+    echo -e '\e[1;31mFAIL\e[0m'
+    exit 1
+fi
+if ! cat $TMPDIR/output.txt | grep "^UNITTEST" | cut -f 2 | $DIFF - $TMPDIR/golden_output.txt ; then
     echo -e '\e[1;31mFAIL\e[0m'
     exit 1
 fi
